Building the skeleton for the builder and dynamic linkage
Builder: for building the charts from scottplot and adjusting the data from AlphaVantage to meet the requirements of the plots.
Dynamic Linkage: a way for loading in ML models. Two main types, the forecasting and price prediction. This was chosen as I have plans for Python.net to load in dll files of ML models trained with TensorFlow.